add bill feature

+++ backend/src/main/java/com/kyocoolcool/keycloak/backend/product/ProductRepository.java
@@ -1,10 +1,9 @@
 package com.kyocoolcool.keycloak.backend.product;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
     List<Product> findProductsByName(String name);
 }
